requirements.txt
# Основные зависимости для работы бота
aiogram==3.22.0
python-dotenv==1.2.1

# Работа с базой данных
sqlite3==3.51.0

# Работа с API
openai==2.8.1